crossdock - Move client into crossdock-go
